Keep test doubles out of conftest: MessageBoxes moves to tests/support.py
conftest.py holds fixtures only, by its own docstring; the recorder the
message_boxes fixture installs is a test double like the others in
tests/support.py.
